rc - refactored helpers files to move shared helper functions into a single file, and adjusted tests to account for that
