+++ javascript/src/main/components/RoomSlots/RoomSlotForm.js
@@ -1,4 +1,4 @@
 import React, { useState, useRef } from "react";
 import { Form, Button, Row, Col } from "react-bootstrap";
-import { checkTime } from "main/utils/OfficeHourFormHelpers";
+import { checkTime } from "main/utils/FormHelpers";
 import TutorAssignmentSelect from "main/components/TutorAssignment/TutorAssignmentSelect";
