+++ javascript/src/test/utils/CourseFormHelpers.test.js
@@ -1,21 +1,6 @@
-import { checkCourseQuarter, checkEmail } from "main/utils/CourseFormHelpers";
+import { checkEmail } from "main/utils/CourseFormHelpers";
 
 describe("course form helpers test", () => {
-    describe("checkCourseQuarter", () => {
-        test("good quarters", () => {
-            expect(checkCourseQuarter("F20")).toBe(true);
-            expect(checkCourseQuarter("W20")).toBe(true);
-            expect(checkCourseQuarter("S20")).toBe(true);
-            expect(checkCourseQuarter("M20")).toBe(true);
-        });
-
-        test("bad quarters", () => {
-            expect(checkCourseQuarter("X20")).toBe(false);
-            expect(checkCourseQuarter("FFF")).toBe(false);
-            expect(checkCourseQuarter("S2020")).toBe(false);
-            expect(checkCourseQuarter("Summer20")).toBe(false);
-        });
-    });
 
     describe("checkEmail", () => {
         test("good email", async () => {
